vajra/backend/requirements.txt
fastapi
uvicorn[standard]
duckdb
google-generativeai
sse-starlette
python-multipart
python-dotenv
pydantic
